Update Bots Folder (#210)
